Add files via upload
The Kaggle dataset can be found here:

https://www.kaggle.com/nicapotato/womens-ecommerce-clothing-reviews

This is a labelled dataset through which we need to predict whether a woman would be likely to recommend or not recommend clothing products. The dataset is mixed- contains text data in the form of reviews and also has other numeric and categorical parameters on which the model can be trained.

I have used two approaches- one just with the reviews and one with both reviews and other categorical and numerical features. When comparing accuracy scores both approaches provide similar range of accuracies when using a Multinomial Naive Bayes algorithm. However when using the combined approach of text + numerical + categorical data the ROC AUC curve score shows significantly lower results than the pure text approach.