Add files via upload
Haptik is one of the world's largest conversational AI platforms. It is a personal assistant mobile app, powered by a combination of artificial intelligence and human assistance. It has its domain in multiple fields including customer support, feedback, order status and live chat.

this is a dataset from Haptik containing the messages it receives from the customers and which topic(class) the messages refer to.

The task is to assign the right class or domain that this message belongs to.

This is a Multi Class Classification problem statement.